+++ all-is-cubes/src/block.rs
@@ -11,7 +11,7 @@ use alloc::sync::Arc;
 use alloc::vec::Vec;
 use core::fmt;
 
-use crate::inv::{self, InvInBlock};
+use crate::inv;
 use crate::listen::{Listen as _, Listener};
 use crate::math::{GridAab, GridCoordinate, GridPoint, GridRotation, GridVector, Rgb, Rgba, Vol};
 use crate::space::{SetCubeError, Space, SpaceChange};
@@ -354,18 +354,25 @@ impl Block {
 
     /// Given a block that does not yet have an [`Modifier::Inventory`], add it.
     ///
-    /// The size of the added inventory is the maximum of the size set by [`InvInBlock`]
+    /// The size of the added inventory is the maximum of the size set by [`config`]
+    /// (which should normally be from the evaluated [`BlockAttributes`] of this block)
     /// and the size of `contents`.
-    ///
-    /// TODO: It is not yet decided what happens when this is called on a block which already
-    /// has an inventory. Currently, it just attaches another modifier.
+    //---
+    // TODO(inventory): Decide what happens when this is called on a block which
+    // already has an inventory. Currently, it just attaches another modifier.
+    //
+    // TODO(inventory): Decide what happens when `config.size == 0`.
+    // Should we refrain from adding the modifier?
     #[must_use]
-    pub fn with_inventory(self, contents: impl Iterator<Item = inv::Slot>) -> Block {
+    pub fn with_inventory(
+        self,
+        config: &inv::InvInBlock,
+        contents: impl Iterator<Item = inv::Slot>,
+    ) -> Block {
         let inventory = inv::Inventory::from_slots(
             itertools::Itertools::zip_longest(
                 contents,
-                // TODO(inventory): InvInBlock should be part of block attributes
-                core::iter::repeat(inv::Slot::Empty).take(InvInBlock::default().size),
+                core::iter::repeat(inv::Slot::Empty).take(config.size),
             )
             .map(|z| z.into_left())
             .collect::<Box<[inv::Slot]>>(),