+++ all-is-cubes/src/inv.rs
@@ -8,7 +8,7 @@ pub use icons::*;
 mod inventory;
 pub use inventory::*;
 mod inv_in_block;
-pub use inv_in_block::InvInBlock;
+pub use inv_in_block::*;
 mod tool;
 pub use tool::*;