+++ all-is-cubes/src/inv/inv_in_block.rs
@@ -3,15 +3,17 @@
 use alloc::vec::Vec;
 
 use crate::block::Resolution;
-use crate::math::{GridCoordinate, GridPoint, GridVector};
+use crate::math::{GridCoordinate, GridPoint, GridRotation, GridVector, Gridgid};
 
 #[cfg(doc)]
 use crate::block::Modifier;
 
 /// Defines how a [`Modifier::Inventory`] should be configured and displayed within a block.
-///
-/// TODO: Needs a better name.
-#[derive(Clone, Debug, Eq, PartialEq)]
+//---
+// TODO(inventory): better name?
+// TODO(inventory): needs accessors or public fields
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
+#[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]
 pub struct InvInBlock {
     /// Number of slots the inventory should have.
     pub(crate) size: usize, // TODO: use platform-independent max size
@@ -27,10 +29,15 @@ pub struct InvInBlock {
     /// beyond this resolution.
     pub(crate) icon_resolution: Resolution,
 
+    // TODO: following the rule for all `BlockAttributes` being cheap to clone,
+    // this should be `Arc<[IconRow]>`, but that's mildly inconvenient for `Arbitrary`, so
+    // I'm not bothering for this first iteration.
+    // (But maybe we should be handling that at a higher level of the structure.)
     pub(crate) icon_rows: Vec<IconRow>,
 }
 
-#[derive(Clone, Debug, Eq, PartialEq)]
+/// Positioning of a displayed row of inventory icons; part of [`InvInBlock`].
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
 pub struct IconRow {
     pub(crate) first_slot: usize,
     pub(crate) count: usize,
@@ -39,25 +46,17 @@ pub struct IconRow {
 }
 
 impl InvInBlock {
-    /// Returns which inventory slots should be rendered as icons, and the lower corners
-    /// of the icons.
-    pub(crate) fn icon_positions(&self) -> impl Iterator<Item = (usize, GridPoint)> + '_ {
-        self.icon_rows.iter().flat_map(|row| {
-            (0..row.count).map_while(move |sub_index| {
-                let slot_index = row.first_slot.checked_add(sub_index)?;
-                Some((
-                    slot_index,
-                    // TODO: this should be checked arithmetic
-                    row.origin + row.stride * GridCoordinate::try_from(sub_index).ok()?,
-                ))
-            })
-        })
-    }
-}
+    /// Value for blocks which should not carry inventories.
+    pub(crate) const EMPTY: Self = Self {
+        size: 0,
+        icon_scale: Resolution::R1,      // arbitrary
+        icon_resolution: Resolution::R1, // arbitrary
+        icon_rows: Vec::new(),
+    };
 
-impl Default for InvInBlock {
-    fn default() -> Self {
-        // TODO: placeholder; the real default should be empty (zero slots, invisible).
+    // TODO(inventory): this substitutes for the lack of ability to actually construct one publicly
+    #[doc(hidden)]
+    pub fn new_placeholder() -> Self {
         Self {
             size: 1,
             icon_scale: Resolution::R4,
@@ -84,6 +83,122 @@ impl Default for InvInBlock {
             ],
         }
     }
+
+    /// Returns which inventory slots should be rendered as icons, and the lower corners
+    /// of the icons.
+    pub(crate) fn icon_positions(&self) -> impl Iterator<Item = (usize, GridPoint)> + '_ {
+        self.icon_rows.iter().flat_map(|row| {
+            (0..row.count).map_while(move |sub_index| {
+                let slot_index = row.first_slot.checked_add(sub_index)?;
+                Some((
+                    slot_index,
+                    // TODO: this should be checked arithmetic
+                    row.origin + row.stride * GridCoordinate::try_from(sub_index).ok()?,
+                ))
+            })
+        })
+    }
+
+    pub(crate) fn rotationally_symmetric(&self) -> bool {
+        // If it doesn't display any icons, then it's symmetric.
+        self.icon_rows.is_empty()
+    }
+
+    pub(crate) fn rotate(self, rotation: GridRotation) -> Self {
+        let Self {
+            size,
+            icon_scale,
+            icon_resolution,
+            icon_rows,
+        } = self;
+        let transform = rotation.to_positive_octant_transform(icon_resolution.into());
+        Self {
+            size,
+            icon_scale,
+            icon_resolution,
+            icon_rows: icon_rows
+                .into_iter()
+                .map(|row| row.rotate(transform))
+                .collect(),
+        }
+    }
+
+    /// Combine the two inputs to form one which has the size and display of both.
+    pub(crate) fn concatenate(self, other: InvInBlock) -> InvInBlock {
+        if self.size == 0 {
+            other
+        } else {
+            Self {
+                size: self.size.saturating_add(other.size),
+                // TODO(inventory): scale and resolution need adaptation
+                icon_scale: self.icon_scale,
+                icon_resolution: self.icon_resolution,
+                icon_rows: self
+                    .icon_rows
+                    .into_iter()
+                    .chain(other.icon_rows.into_iter().filter_map(|mut row| {
+                        row.first_slot = row.first_slot.checked_add(self.size)?;
+                        Some(row)
+                    }))
+                    .collect(),
+            }
+        }
+    }
+}
+
+impl Default for InvInBlock {
+    /// Returns [`InvInBlock::EMPTY`].
+    fn default() -> Self {
+        Self::EMPTY
+    }
+}
+
+impl crate::universe::VisitHandles for InvInBlock {
+    fn visit_handles(&self, _: &mut dyn crate::universe::HandleVisitor) {
+        let Self {
+            size: _,
+            icon_scale: _,
+            icon_resolution: _,
+            icon_rows: _,
+        } = self;
+    }
+}
+
+impl IconRow {
+    fn rotate(self, transform: Gridgid) -> Self {
+        // TODO: The icons themselves (not their positions) need to be rotated,
+        // but this is not supported yet.
+        Self {
+            first_slot: self.first_slot,
+            count: self.count,
+            origin: transform.transform_point(self.origin), // TODO: does not account for size of icon
+            stride: transform.rotation.transform_vector(self.stride),
+        }
+    }
+}
+
+#[cfg(feature = "arbitrary")]
+impl<'a> arbitrary::Arbitrary<'a> for IconRow {
+    fn arbitrary(u: &mut arbitrary::Unstructured<'a>) -> arbitrary::Result<Self> {
+        // TODO: there are lots of extremely useless out-of-bounds values here;
+        // bias to useful ones?
+        Ok(Self {
+            first_slot: u.arbitrary()?,
+            count: u.arbitrary()?,
+            origin: <[i32; 3]>::arbitrary(u)?.into(),
+            stride: <[i32; 3]>::arbitrary(u)?.into(),
+        })
+    }
+
+    fn size_hint(depth: usize) -> (usize, Option<usize>) {
+        use arbitrary::{size_hint::and_all, Arbitrary};
+        and_all(&[
+            <usize as Arbitrary>::size_hint(depth),
+            <usize as Arbitrary>::size_hint(depth),
+            <[GridCoordinate; 3] as Arbitrary>::size_hint(depth),
+            <[GridCoordinate; 3] as Arbitrary>::size_hint(depth),
+        ])
+    }
 }
 
 #[cfg(test)]
@@ -92,10 +207,10 @@ mod tests {
     use euclid::point3;
     use pretty_assertions::assert_eq;
 
-    // TODO: this test should be revised to create an `InvInBlock` for testing instead of the default.
     #[test]
-    fn default_icon_positions() {
-        let iib = InvInBlock::default();
+    fn icon_positions_output() {
+        // TODO: this test should be revised to create a specific `InvInBlock` value for testingg
+        let iib = InvInBlock::new_placeholder();
         assert_eq!(
             iib.icon_positions().take(10).collect::<Vec<_>>(),
             vec![